backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
openai==1.68.0
browser-use==1.0.0
playwright==1.46.0
websockets==13.1
aiohttp==3.10.8
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
asyncio-throttle==1.0.2
tenacity==9.0.0

